Big Changes, implemented the template pattern
I implemented the template pattern here by making a skeleton of what I want my roulette game to follow. I made methods that my RouletteGame will extend from this template so then in main when we call playGame() from RouletteGameTemplate the Roulette game must follow this pattern or sequence of events. I had to seperate my big play() into seperate functions so that we can do this but. The code is the same however I am just seperating it so that we can follow this template pattern.