Made a multiplier for the bets
Here I made a multipler for the bets. I did this by making a getmultipler() method that takes in a parameter of betType and has a switch statement going through each case or in our case the assoiciated bet. Then we return a number which is the multipler and then in the updateChips method we multiply that number by the betAmount. Then I just updated the parameters for updateChips